Slider

+++ ProniaTask/ProniaTask.Business/Services/Concretes/SliderService.cs
@@ -1,4 +1,6 @@
-﻿using ProniaTask.Business.Services.Abstracts;
+﻿using ProniaTask.Business.Enums;
+using ProniaTask.Business.Exceptions;
+using ProniaTask.Business.Services.Abstracts;
 using ProniaTask.Core.Models;
 using ProniaTask.Core.RepositoryAbstracts;
 using System;
@@ -20,11 +22,15 @@ public class SliderService : ISliderService
 
     public async Task AddSlider(Slider slider)
     {
+        if (slider.ImageFile.ContentType != "image/png" && slider.ImageFile.ContentType != "image/jpeg")
+            throw new ImageContentTypeException("Fayl formati duzgun deyil!");
+        if (slider.ImageFile.Length > 2097152)
+            throw new ImageSizeException("Sheklin olcusu max 2mb ola biler");
 
         var extension = Path.GetExtension(slider.ImageFile.FileName);
         var fileName = $"slider={Guid.NewGuid().ToString().ToLower()}{extension}";
          
-        string path = "C:\\Users\\memme\\Desktop\\MVC---N-Layer-Generic-Repository-main\\ProniaTask\\ProniaTask\\wwwroot\\admin\\uploads\\sliders\\" + fileName ;
+        string path = "C:\\Users\\memme\\Desktop\\MVC---N-Layer-Generic-Repository\\ProniaTask\\ProniaTask\\wwwroot\\uploads\\sliders\\" + fileName ;
 
 ;       using (FileStream fileStream = new FileStream(path, FileMode.Create))
         {
@@ -40,7 +46,8 @@ public class SliderService : ISliderService
     public void DeleteSlider(int id)
     {
         var existSlider = _sliderRepository.Get(x => x.Id == id);
-       _sliderRepository.Delete(existSlider);
+        if (existSlider == null) throw new EntityNotFoundException("Bele slider yoxdur!");
+        _sliderRepository.Delete(existSlider);
         _sliderRepository.Commit();
     }
 
@@ -54,17 +61,34 @@ public class SliderService : ISliderService
         return _sliderRepository.Get(func);
     }
 
-    public void UpdateSlider(int id, Slider newSliders)
+    public void UpdateSlider(int id, Slider newSlider)
     {
-        var oldSlider = _sliderRepository.Get(x=>x.Id == id);
+        var oldSlider = _sliderRepository.Get(x => x.Id == id);
+        if (oldSlider == null) throw new EntityNotFoundException("Bele bir slider yoxdur!");
 
-        oldSlider.Offer = newSliders.Offer;
-        oldSlider.Title = newSliders.Title;
-        oldSlider.Description = newSliders.Description;
-        oldSlider.RedirectUrl = newSliders.RedirectUrl;
-        
+		if (newSlider.ImageFile.ContentType != "image/png" && newSlider.ImageFile.ContentType != "image/jpeg")
+			throw new ImageContentTypeException("Fayl formati duzgun deyil!");
+		if (newSlider.ImageFile.Length > 2097152)
+			throw new ImageSizeException("Sheklin olcusu max 2mb ola biler");
 
-        _sliderRepository.Commit();
+		var extension = Path.GetExtension(newSlider.ImageFile.FileName);
+		var fileName = $"slider={Guid.NewGuid().ToString().ToLower()}{extension}";
+
+		string path = "C:\\Users\\memme\\Desktop\\MVC---N-Layer-Generic-Repository\\ProniaTask\\ProniaTask\\wwwroot\\uploads\\sliders\\" + fileName;
+
+		; using (FileStream fileStream = new FileStream(path, FileMode.Create))
+		{
+			newSlider.ImageFile.CopyTo(fileStream);
+		}
+
+		oldSlider.ImageUrl = fileName;
+        oldSlider.Title = newSlider.Title;
+        oldSlider.Description = newSlider.Description;
+        oldSlider.Offer = newSlider.Description;
+
+
+
+		_sliderRepository.Commit();
        
     }
 }
